Fix sign conversions in gameplay and coordinate code
Fix types at their source across both games' simulation and map code:
target values, the stage counter, radar origin, unit and building limits,
MaxStrength (signed at the same width) and TD's building capacity become
signed; prerequisite and team-type masks become uint64_t, which also makes
1 << btype defined past bit 31; one-bit flags are assigned as bools.

Packing a signed component into a COORDINATE, TARGET or LEPTON, or reading
one back, is now an explicit cast at that point, and TD's Coord_Add and
Coord_Sub no longer pun through short pointers. Serialized fields keep their
widths, and the RA and TD headless save/load checks pass. Places that pass
a CELL where a COORDINATE or TARGET is expected are marked Suspicious and
left unchanged.

+++ src/ra/ccini.cc
@@ -1494,7 +1494,7 @@ int CCINIClass::Get_Unique_ID() const {
     ((CCINIClass*)this)->Calculate_Message_Digest();
   }
 
-  return CrcEngine::Compute(Digest);
+  return static_cast<int>(CrcEngine::Compute(Digest));
 }
 
 /***********************************************************************************************
